+++ src/API.php
@@ -11,25 +11,30 @@ namespace Geniem\RediPress;
  * @param mixed $post_id The ID to fetch.
  * @return WP_Post|null The post to fetch or null if not found.
  */
-function get_post( $post_id ) : ?\WP_Post {
+function get_post( $post_id ) { //}: ?\WP_Post {
     $settings = new Settings();
-    $client = apply_filters( 'redipress/client', null );
+    $client   = apply_filters( 'redipress/client', null );
 
     $index = Settings::get( 'index' );
 
     $result = $client->raw_command( 'FT.GET', [ $index, $post_id ] );
+
+    // If nothing is found, just return null
+    if ( ! $result ) {
+        return null;
+    }
 
     $result = Utility::format( $result );
 
     if ( ! empty( $result['post_object'] ) ) {
         $post_object = maybe_unserialize( $result['post_object'] );
 
-        if ( $post_object instanceof WP_Post ) {
+        if ( $post_object instanceof \WP_Post ) {
             return $post_object;
         }
     }
 
-    return null;
+    return false;
 }
 
 /**